[e2e] move test helpers from e2e to utils/common
