+++ test/new-e2e/pkg/utils/e2e/client/agent_client.go
@@ -19,7 +19,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/DataDog/datadog-agent/test/new-e2e/pkg/e2e"
+	"github.com/DataDog/datadog-agent/test/new-e2e/pkg/runner"
 	"github.com/DataDog/datadog-agent/test/new-e2e/pkg/utils/common"
 	"github.com/DataDog/datadog-agent/test/new-e2e/pkg/utils/e2e/client/agentclient"
 	"github.com/DataDog/datadog-agent/test/new-e2e/pkg/utils/e2e/client/agentclientparams"
@@ -197,11 +197,15 @@ func waitForReadyTimeout(t *testing.T, host *Host, commandRunner *agentCommandRu
 }
 
 func generateAndDownloadFlare(t *testing.T, commandRunner *agentCommandRunner, host *Host) error {
-	root, err := e2e.CreateRootOutputDir()
+	outputRoot, err := runner.GetProfile().GetOutputDir()
+	if err != nil {
+		return err
+	}
+	root, err := common.CreateRootOutputDir(outputRoot)
 	if err != nil {
 		return fmt.Errorf("could not get root output directory: %w", err)
 	}
-	outputDir, err := e2e.CreateTestOutputDir(root, t)
+	outputDir, err := common.CreateTestOutputDir(root, t)
 	if err != nil {
 		return fmt.Errorf("could not get output directory: %w", err)
 	}
